refactor move activateButton into own file
